requirements.txt
python-dotenv
selenium
fastapi
uvicorn[standard]
